using double instead of float

+++ cytopy/util/simdata.py
@@ -37,8 +37,8 @@ def simdata(N=[300, 100, 200], J=25, a_W=[200., 500., 200., 100.], L0=5, L1=3, s
     if sig is None:
         sig = torch.ones((I, )) * 1.0
 
-    mu0 = -(torch.arange(L0) + 3 * sig.max()).float()
-    mu1 = (torch.arange(L1) + 3 * sig.max()).float()
+    mu0 = -(torch.arange(L0) + 3 * sig.max()).double()
+    mu1 = (torch.arange(L1) + 3 * sig.max()).double()
 
     params = {'W': W, 'v': v, 'eta0': eta0, 'eta1': eta1,
               'mu0': mu0, 'mu1': mu1, 'sig': sig, 'Z': Z}
